Use unique board ID in LocalBoardRO

+++ Detectors/MUON/MID/Raw/include/MIDRaw/ColumnDataToLocalBoard.h
@@ -39,7 +39,7 @@ class ColumnDataToLocalBoard
 
  private:
   bool keepBoard(const LocalBoardRO& loc) const;
-  std::unordered_map<uint16_t, LocalBoardRO> mLocalBoardsMap{};      /// Map of data per board
+  std::unordered_map<uint8_t, LocalBoardRO> mLocalBoardsMap{};       /// Map of data per board
   std::unordered_map<uint16_t, std::vector<LocalBoardRO>> mGBTMap{}; /// Map of data per GBT link
   CrateMapper mCrateMapper{};                                        /// Crate mapper
   Mapping mMapping{};                                                /// Segmentation
